move authorization access_map to a model concern to match navigation usage
