+++ app/models/concerns/fae/authorization_concern.rb
@@ -1,0 +1,12 @@
+module Fae
+  module AuthorizationConcern
+    extend ActiveSupport::Concern
+    module ClassMethods
+
+      def access_map
+        {}
+      end
+
+    end
+  end
+end
