+++ app/models/fae/authorization.rb
@@ -1,0 +1,8 @@
+module Fae
+  class Authorization
+    include Fae::AuthorizationConcern
+
+    # `self.access_map` is defined in Fae::AuthorizationConcern on the parent app
+
+  end
+end
