--- spec/dummy/app/controllers/concerns/fae/authorization_access_map.rb
@@ -1,18 +1,0 @@
-module Fae
-  module AuthorizationAccessMap
-    extend ActiveSupport::Concern
-
-    def self.access_map
-      {
-        'people' => ['super admin', 'admin'],
-        'locations' => ['super admin', 'admin'],
-        'validation_testers' => ['super admin', 'admin'],
-        'releases' => ['super admin', 'admin'],
-        'selling_points' => ['super admin', 'admin'],
-        'jerseys' => ['super admin', 'admin'],
-        'content_blocks/about_us' => ['super admin']
-      }
-    end
-
-  end
-end
